+++ frontend/src/containers/liquidity/LiquidityList.js
@@ -6,8 +6,7 @@ import Input from '../../components/shared/Input';
 import InputToken from '../../components/shared/InputToken';
 import ButtonDivider from '../../components/shared/ButtonDivider';
 import Button from '../../components/shared/Button';
-import cryptoCurrencies from '../../constants/cryptoCurrencies';
-import pairTokens from '../../constants/pairTokens';
+import cryptoCurrencies from '../../constants/tokens';
 import TokenPair from './TokenPair';
 import {PactContext} from '../../contexts/PactContext'
 import {reduceBalance} from '../../utils/reduceBalance';