+++ frontend/src/contexts/PactContext.js
@@ -5,42 +5,40 @@ import CryptoJS from 'crypto-js'
 import { NotificationContext, STATUSES } from './NotificationContext';
 import PasswordPopup from '../components/shared/PasswordPopup';
 import { toast } from 'react-toastify';
-import pairTokens from '../constants/pairTokens'
+import pairTokens from '../constants/pairs.json'
 import swal from '@sweetalert/with-react'
 import pwPrompt from '../components/alerts/pwPrompt'
 import walletError from '../components/alerts/walletError'
 import walletSigError from '../components/alerts/walletSigError'
 import walletLoading from '../components/alerts/walletLoading'
-import tokenData from '../constants/cryptoCurrencies'
-import { reduceBalance, keepDecimal, extractDecimal } from '../utils/reduceBalance'
+import { reduceBalance, extractDecimal } from '../utils/reduceBalance'
+import tokenData from '../constants/cryptoCurrencies';
 
 export const PactContext = createContext();
-
 const savedAcct = localStorage.getItem('acct');
 const savedPrivKey = localStorage.getItem('pk');
 const savedNetwork = localStorage.getItem('network');
 const savedSlippage = localStorage.getItem('slippage');
 const savedSigning = localStorage.getItem('signing');
 const savedTtl = localStorage.getItem('ttl');
-
-const network = "https://us1.testnet.chainweb.com/chainweb/0.0/testnet04/chain/0/pact";
-
 const chainId = "0";
+const PRECISION = 12;
+const NETWORKID = 'testnet04';
+const network = `https://us1.testnet.chainweb.com/chainweb/0.0/${NETWORKID}/chain/${chainId}/pact`;
+
 const creationTime = () => Math.round((new Date).getTime()/1000)-10;
-const GAS_PRICE = 0.000000000001
-const PRECISION = 12
+const GAS_PRICE = 0.000000000001;
 
 export const PactProvider = (props) => {
-
   const notificationContext = useContext(NotificationContext);
-
   const [account, setAccount] = useState((savedAcct ? JSON.parse(savedAcct) : {account: null, guard: null, balance: 0}));
   const [tokenAccount, setTokenAccount] = useState({account: null, guard: null, balance: 0});
   const [privKey, setPrivKey] = useState((savedPrivKey ? savedPrivKey : ""));
   const keyPair = privKey ? Pact.crypto.restoreKeyPairFromSecretKey(privKey) : "";
   const [tokenFromAccount, setTokenFromAccount] = useState({account: null, guard: null, balance: 0});
   const [tokenToAccount, setTokenToAccount] = useState({account: null, guard: null, balance: 0});
-  const [tokenList, setTokenList] = useState({tokens: []});
+  const [tokenList, setTokenList] = useState(tokenData);
+  const [precision, setPrecision] = useState(false);
   const [pairAccount, setPairAccount] = useState("");
   const [pairReserve, setPairReserve] = useState("");
   const [pair, setPair] = useState("");
@@ -68,7 +66,6 @@ export const PactProvider = (props) => {
   //TO FIX, not working when multiple toasts are there
   const toastId = React.useRef(null)
   // const [toastIds, setToastIds] = useState({})
-
   useEffect(() => {
     if (account.account) setRegistered(true);
   }, [registered]);
@@ -82,24 +79,56 @@ export const PactProvider = (props) => {
   }, [sendRes])
 
   useEffect(() => {
-    fetchAllBalances();
-  }, [balances, account.account, sendRes])
-
-  useEffect(() => {
     const store = async () => localStorage.setItem('signing', JSON.stringify(signing));
     store()
   }, [signing])
 
+  useEffect(() => {
+    fetchPrecision();
+  }, [precision])
+
+  useEffect(() => {
+    fetchAllBalances();
+  }, [balances, account.account, sendRes])
+
+  const pollingNotif = (reqKey) => {
+    return (
+      toastId.current = notificationContext.showNotification({
+              title: 'Transaction Pending',
+              message: reqKey,
+              type: STATUSES.INFO,
+              autoClose: 92000,
+              hideProgressBar: false
+            }
+      )
+    )
+  }
+
+  const getCorrectBalance = (balance) => {
+    const balanceClean = (!isNaN(balance) ? balance : balance.decimal)
+    return balanceClean
+  }
+
+  const storeSlippage = async (slippage) => {
+    await setSlippage(slippage)
+    await localStorage.setItem('slippage', slippage);
+  }
+
+  const storeTtl = async (ttl) => {
+    await setTtl(slippage)
+    await localStorage.setItem('ttl', ttl);
+  }
+
   const fetchAllBalances = async () => {
     let count=0;
     let endBracket = ''
-    let tokenNames = Object.values(tokenData).reduce((accum, cum)=> {
+    let tokenNames = Object.values(tokenData).reduce((accum, cumul)=> {
       count++;
       endBracket+=')'
       let code =  `
       (let
-        ((${cum.code}
-          (try -1 (${cum.name}.get-balance "${account.account}"))
+        ((${cumul.name}
+          (try -1 (${cumul.code}.get-balance "${account.account}"))
       ))`
       accum+=code;
       return accum;
@@ -127,34 +156,38 @@ export const PactProvider = (props) => {
     }
   }
 
+  const fetchPrecision = async () => {
+    let count=0;
+    let endBracket = ''
+    let tokenNames = Object.values(tokenData).reduce((accum, cumul)=> {
+      count++;
+      endBracket+=')'
+      let code =  `
+      (let
+        ((${cumul.name}
+          (try -1 (${cumul.code}.precision))
+      ))`
+      accum+=code;
+      return accum;
+    }, '')
+    let objFormat =  `{${Object.keys(tokenData).map(token => `"${token}": ${token}`).join(',')}}`
+    tokenNames = tokenNames + objFormat + endBracket;
+    try {
+      let data = await Pact.fetch.local({
+          pactCode: tokenNames,
+          meta: Pact.lang.mkMeta("", chainId ,GAS_PRICE,3000,creationTime(), 600),
+        }, network);
+      if (data.result.status === "success"){
+        Object.keys(tokenData).forEach(token => {
+          tokenData[token].precision = extractDecimal(data.result.data[token]);
+        })
+        setPrecision(true);
+      }
+    } catch (e) {
+      setPrecision(false);
 
-  const pollingNotif = (reqKey) => {
-    return (
-      toastId.current = notificationContext.showNotification({
-              title: 'Transaction Pending',
-              message: reqKey,
-              type: STATUSES.INFO,
-              autoClose: 92000,
-              hideProgressBar: false
-            }
-      )
-    )
-  }
-
-
-  const getCorrectBalance = (balance) => {
-    const balanceClean = (!isNaN(balance) ? balance : balance.decimal)
-    return balanceClean
-  }
-
-  const storeSlippage = async (slippage) => {
-    await setSlippage(slippage)
-    await localStorage.setItem('slippage', slippage);
-  }
-
-  const storeTtl = async (ttl) => {
-    await setTtl(slippage)
-    await localStorage.setItem('ttl', ttl);
+      console.log(e)
+    }
   }
 
   const setVerifiedAccount = async (accountName) => {
@@ -220,7 +253,7 @@ export const PactProvider = (props) => {
               ""
             )`,
           meta: Pact.lang.mkMeta("", chainId ,GAS_PRICE,5000,creationTime(),28800),
-          networkId: "testnet04"
+          networkId: NETWORKID
       }, network);
       let pair =  data.result.data.account
       try {
@@ -234,10 +267,10 @@ export const PactProvider = (props) => {
             (kswap.exchange.add-liquidity
                 ${token0}
                 ${token1}
-                ${keepDecimal(amountDesired0)}
-                ${keepDecimal(amountDesired1)}
-                ${keepDecimal(reduceBalance(amountDesired0*(1-parseFloat(slippage)),PRECISION))}
-                ${keepDecimal(reduceBalance(amountDesired1*(1-parseFloat(slippage)),PRECISION))}
+                (read-decimal 'amountDesired0)
+                (read-decimal 'amountDesired1)
+                (read-decimal 'amountMinimum0)
+                (read-decimal 'amountMinimum1)
                 ${JSON.stringify(account.account)}
                 ${JSON.stringify(account.account)}
                 (read-keyset 'user-ks)
@@ -251,10 +284,14 @@ export const PactProvider = (props) => {
               ]
             },
             envData: {
-              "user-ks": [keyPair.publicKey]
+              "user-ks": [keyPair.publicKey],
+              "amountDesired0": amountDesired0,
+              "amountDesired1": amountDesired1,
+              "amountMinimum0": reduceBalance(amountDesired0*(1-parseFloat(slippage)),tokenData[token0].precision),
+              "amountMinimum1": reduceBalance(amountDesired1*(1-parseFloat(slippage)),tokenData[token1].precision)
             },
             meta: Pact.lang.mkMeta("kswap-free-gas", chainId ,GAS_PRICE,5000,creationTime(), 600),
-            networkId: "testnet04"
+            networkId: NETWORKID
           };
         data = await Pact.fetch.local(cmd, network);
         setCmd(cmd);
@@ -277,15 +314,15 @@ export const PactProvider = (props) => {
       if (privKey.length !== 64) {
         return
       }
-      let pair = await getPairAccount(token0, token1);
+      let pair = await getPairAccount(token0.code, token1.code);
       let cmd = {
           pactCode: `(kswap.exchange.add-liquidity
-              ${token0}
-              ${token1}
-              ${keepDecimal(amountDesired0)}
-              ${keepDecimal(amountDesired1)}
-              ${keepDecimal(reduceBalance(amountDesired0*(1-parseFloat(slippage)), PRECISION))}
-              ${keepDecimal(reduceBalance(amountDesired1*(1-parseFloat(slippage)), PRECISION))}
+              ${token0.code}
+              ${token1.code}
+              (read-decimal 'amountDesired0)
+              (read-decimal 'amountDesired1)
+              (read-decimal 'amountMinimum0)
+              (read-decimal 'amountMinimum1)
               ${JSON.stringify(account.account)}
               ${JSON.stringify(account.account)}
               (read-keyset 'user-ks)
@@ -294,16 +331,20 @@ export const PactProvider = (props) => {
             publicKey: account.guard.keys[0],
             secretKey: privKey,
             clist: [
-              {name: `${token0}.TRANSFER`, args: [account.account, pair, Number(amountDesired0)]},
-              {name: `${token1}.TRANSFER`, args: [account.account, pair, Number(amountDesired1)]},
+              {name: `${token0.code}.TRANSFER`, args: [account.account, pair, Number(amountDesired0)]},
+              {name: `${token1.code}.TRANSFER`, args: [account.account, pair, Number(amountDesired1)]},
               {name: "kswap.gas-station.GAS_PAYER", args: ["free-gas", {int: 1}, 1.0]},
             ]
           },
           envData: {
-            "user-ks": account.guard
+            "user-ks": account.guard,
+            "amountDesired0": amountDesired0,
+            "amountDesired1": amountDesired1,
+            "amountMinimum0": reduceBalance(amountDesired0*(1-parseFloat(slippage)),tokenData[token0.name].precision),
+            "amountMinimum1": reduceBalance(amountDesired1*(1-parseFloat(slippage)),tokenData[token1.name].precision)
           },
           meta: Pact.lang.mkMeta("kswap-free-gas", chainId ,GAS_PRICE,3000,creationTime(), 600),
-          networkId: "testnet04"
+          networkId: NETWORKID
         };
       let data = await Pact.fetch.local(cmd, network);
       setCmd(cmd);
@@ -311,36 +352,42 @@ export const PactProvider = (props) => {
       return data;
     } catch (e) {
       setLocalRes({});
-      return -1
       console.log(e)
+      return -1
     }
   }
 
   const addLiquidityWallet = async (token0, token1, amountDesired0, amountDesired1) => {
     try {
-      let pair = await getPairAccount(token0, token1);
+      let pair = await getPairAccount(token0.code, token1.code);
       const signCmd = {
         pactCode: `(kswap.exchange.add-liquidity
-            ${token0}
-            ${token1}
-            ${keepDecimal(amountDesired0)}
-            ${keepDecimal(amountDesired1)}
-            ${keepDecimal(reduceBalance(amountDesired0*(1-parseFloat(slippage)),PRECISION))}
-            ${keepDecimal(reduceBalance(amountDesired1*(1-parseFloat(slippage)),PRECISION))}
+            ${token0.code}
+            ${token1.code}
+            (read-decimal 'amountDesired0)
+            (read-decimal 'amountDesired1)
+            (read-decimal 'amountMinimum0)
+            (read-decimal 'amountMinimum1)
             ${JSON.stringify(account.account)}
             ${JSON.stringify(account.account)}
             (read-keyset 'user-ks)
           )`,
         caps: [
           Pact.lang.mkCap("Gas Station", "free gas", "kswap.gas-station.GAS_PAYER", ["free-gas", {int: 1}, 1.0]),
-          Pact.lang.mkCap("transfer capability", "Transfer Token to Pool", `${token0}.TRANSFER`, [account.account, pair, Number(amountDesired0)]),
-          Pact.lang.mkCap("transfer capability", "Transfer Token to Pool", `${token1}.TRANSFER`, [account.account, pair, Number(amountDesired1)]),
+          Pact.lang.mkCap("transfer capability", "Transfer Token to Pool", `${token0.code}.TRANSFER`, [account.account, pair, Number(amountDesired0)]),
+          Pact.lang.mkCap("transfer capability", "Transfer Token to Pool", `${token1.code}.TRANSFER`, [account.account, pair, Number(amountDesired1)]),
         ],
         sender: "kswap-free-gas",
         gasLimit: 3000,
         chainId: chainId,
         ttl: 600,
-        envData: { "user-ks": account.guard }
+        envData: {
+          "user-ks": account.guard,
+          "amountDesired0": amountDesired0,
+          "amountDesired1": amountDesired1,
+          "amountMinimum0": reduceBalance(amountDesired0*(1-parseFloat(slippage)),tokenData[token0.name].precision),
+          "amountMinimum1": reduceBalance(amountDesired1*(1-parseFloat(slippage)),tokenData[token1.name].precision)
+        }
       }
       //alert to sign tx
       walletLoading();
@@ -374,20 +421,19 @@ export const PactProvider = (props) => {
         return
       }
       let pairKey = await getPairKey(token0, token1);
-      liquidity = keepDecimal(liquidity);
       let pair = await getPairAccount(token0, token1);
       let cmd = {
           pactCode: `(kswap.exchange.remove-liquidity
               ${token0}
               ${token1}
-              ${liquidity}
+              (read-decimal 'liquidity)
               0.0
               0.0
               ${JSON.stringify(account.account)}
               ${JSON.stringify(account.account)}
               (read-keyset 'user-ks)
             )`,
-            networkId: "testnet04",
+            networkId: NETWORKID,
           keyPairs: {
             publicKey: account.guard.keys[0],
             secretKey: privKey,
@@ -398,7 +444,8 @@ export const PactProvider = (props) => {
             ]
           },
           envData: {
-            "user-ks": account.guard
+            "user-ks": account.guard,
+            "liquidity": liquidity
           },
           meta: Pact.lang.mkMeta("kswap-free-gas", chainId ,GAS_PRICE,3000,creationTime(), 600),
         };
@@ -416,13 +463,12 @@ export const PactProvider = (props) => {
   const removeLiquidityWallet = async (token0, token1, liquidity) => {
     try {
       let pairKey = await getPairKey(token0, token1);
-      liquidity = keepDecimal(liquidity);
       let pair = await getPairAccount(token0, token1);
       const signCmd = {
         pactCode:`(kswap.exchange.remove-liquidity
             ${token0}
             ${token1}
-            ${liquidity}
+            (read-decimal 'liquidity)
             0.0
             0.0
             ${JSON.stringify(account.account)}
@@ -438,7 +484,10 @@ export const PactProvider = (props) => {
         gasLimit: 3000,
         chainId: chainId,
         ttl: 600,
-        envData: { "user-ks": account.guard }
+        envData: {
+          "user-ks": account.guard,
+          "liquidity": liquidity
+        }
       }
       //alert to sign tx
       walletLoading();
@@ -531,11 +580,11 @@ export const PactProvider = (props) => {
             (use kswap.exchange)
             (let*
               (
-                (p (get-pair ${pair.token0.name} ${pair.token1.name}))
-                (reserveA (reserve-for p ${pair.token0.name}))
-                (reserveB (reserve-for p ${pair.token1.name}))
-                (totalBal (kswap.tokens.total-supply (kswap.exchange.get-pair-key ${pair.token0.name} ${pair.token1.name})))
-                (acctBal (kswap.tokens.get-balance (kswap.exchange.get-pair-key ${pair.token0.name} ${pair.token1.name}) ${JSON.stringify(account)}))
+                (p (get-pair ${tokenData[pair.token0].code} ${tokenData[pair.token1].code}))
+                (reserveA (reserve-for p ${tokenData[pair.token0].code}))
+                (reserveB (reserve-for p ${tokenData[pair.token1].code}))
+                (totalBal (kswap.tokens.total-supply (kswap.exchange.get-pair-key ${tokenData[pair.token0].code} ${tokenData[pair.token1].code})))
+                (acctBal (kswap.tokens.get-balance (kswap.exchange.get-pair-key ${tokenData[pair.token0].code} ${tokenData[pair.token1].code}) ${JSON.stringify(account)}))
               )[acctBal totalBal reserveA reserveB (* reserveA (/ acctBal totalBal))(* reserveB (/ acctBal totalBal))])
              `,
             meta: Pact.lang.mkMeta("", chainId ,GAS_PRICE,3000,creationTime(), 600),
@@ -563,10 +612,10 @@ export const PactProvider = (props) => {
             (use kswap.exchange)
             (let*
               (
-                (p (get-pair ${pair.token0.name} ${pair.token1.name}))
-                (reserveA (reserve-for p ${pair.token0.name}))
-                (reserveB (reserve-for p ${pair.token1.name}))
-                (totalBal (kswap.tokens.total-supply (kswap.exchange.get-pair-key ${pair.token0.name} ${pair.token1.name})))
+                (p (get-pair ${tokenData[pair.token0].code} ${tokenData[pair.token1].code}))
+                (reserveA (reserve-for p ${tokenData[pair.token0].code}))
+                (reserveB (reserve-for p ${tokenData[pair.token1].code}))
+                (totalBal (kswap.tokens.total-supply (kswap.exchange.get-pair-key ${tokenData[pair.token0].code} ${tokenData[pair.token1].code})))
               )[totalBal reserveA reserveB])
              `,
             meta: Pact.lang.mkMeta("", chainId ,GAS_PRICE,3000,creationTime(), 600),
@@ -659,16 +708,16 @@ export const PactProvider = (props) => {
       let pair = await getPairAccount(token0.address, token1.address);
 
       const inPactCode = `(kswap.exchange.swap-exact-in
-          ${keepDecimal(token0.amount)}
-          ${keepDecimal(reduceBalance(token1.amount*(1-parseFloat(slippage)), PRECISION))}
+          (read-decimal 'token0Amount)
+          (read-decimal 'token1AmountWithSlippage)
           [${token0.address} ${token1.address}]
           ${JSON.stringify(account.account)}
           ${JSON.stringify(account.account)}
           (read-keyset 'user-ks)
         )`
       const outPactCode = `(kswap.exchange.swap-exact-out
-          ${keepDecimal(token1.amount)}
-          ${keepDecimal(reduceBalance(token0.amount*(1+parseFloat(slippage)), PRECISION))}
+          (read-decimal 'token1Amount)
+          (read-decimal 'token0AmountWithSlippage)
           [${token0.address} ${token1.address}]
           ${JSON.stringify(account.account)}
           ${JSON.stringify(account.account)}
@@ -680,15 +729,19 @@ export const PactProvider = (props) => {
             publicKey: account.guard.keys[0],
             secretKey: privKey,
             clist: [
-              {name: `${token0.address}.TRANSFER`, args: [account.account, pair, reduceBalance(token0.amount*(1+parseFloat(slippage)), PRECISION)]},
+              {name: `${token0.address}.TRANSFER`, args: [account.account, pair,  reduceBalance(isSwapIn ? token0.amount : token0.amount*(1+parseFloat(slippage), tokenData[token0.name].precision))]},
             ]
           },
           envData: {
-            "user-ks": account.guard
+            "user-ks": account.guard,
+            "token0Amount": token0.amount,
+            "token1Amount": token1.amount,
+            "token1AmountWithSlippage": reduceBalance(token1.amount*(1-parseFloat(slippage)), tokenData[token1.name].precision),
+            "token0AmountWithSlippage": reduceBalance(token0.amount*(1+parseFloat(slippage)), tokenData[token0.name].precision)
           },
           meta: Pact.lang.mkMeta("", "" ,0,0,0,0),
-          networkId: "testnet04",
-          meta: Pact.lang.mkMeta(account.account, chainId ,GAS_PRICE,3000,creationTime(), 600),
+          networkId: NETWORKID,
+          meta: Pact.lang.mkMeta(account.account, chainId ,GAS_PRICE, 3000, creationTime(), 600),
       }
       setCmd(cmd);
       let data = await Pact.fetch.send(cmd, network);
@@ -710,16 +763,16 @@ export const PactProvider = (props) => {
       const ct = creationTime();
       let pair = await getPairAccount(token0.address, token1.address);
       const inPactCode = `(kswap.exchange.swap-exact-in
-          ${keepDecimal(token0.amount)}
-          ${keepDecimal(reduceBalance(token1.amount*(1-parseFloat(slippage)), PRECISION))}
+          (read-decimal 'token0Amount)
+          (read-decimal 'token1AmountWithSlippage)
           [${token0.address} ${token1.address}]
           ${JSON.stringify(account.account)}
           ${JSON.stringify(account.account)}
           (read-keyset 'user-ks)
         )`
       const outPactCode = `(kswap.exchange.swap-exact-out
-          ${keepDecimal(token1.amount)}
-          ${keepDecimal(reduceBalance(token0.amount*(1+parseFloat(slippage)), PRECISION))}
+          (read-decimal 'token1Amount)
+          (read-decimal 'token0AmountWithSlippage)
           [${token0.address} ${token1.address}]
           ${JSON.stringify(account.account)}
           ${JSON.stringify(account.account)}
@@ -740,15 +793,19 @@ export const PactProvider = (props) => {
                 args: [
                   account.account,
                   pair,
-                  reduceBalance(isSwapIn ? token0.amount : reduceBalance(token0.amount*(1+parseFloat(slippage)), PRECISION)),
+                  reduceBalance(isSwapIn ? token0.amount : token0.amount*(1+parseFloat(slippage)), tokenData[token0.coin].precision),
                 ]
               },
             ]
           },
           envData: {
-            "user-ks": account.guard
+            "user-ks": account.guard,
+            "token0Amount": token0.amount,
+            "token1Amount": token1.amount,
+            "token1AmountWithSlippage": reduceBalance(token1.amount*(1-parseFloat(slippage)), tokenData[token1.coin].precision),
+            "token0AmountWithSlippage": reduceBalance(token0.amount*(1+parseFloat(slippage)), tokenData[token0.coin].precision)
           },
-          networkId: "testnet04",
+          networkId: NETWORKID,
           meta: Pact.lang.mkMeta("kswap-free-gas", chainId, GAS_PRICE, 3000, ct, 600),
       }
       setCmd(cmd);
@@ -766,16 +823,16 @@ export const PactProvider = (props) => {
   const swapWallet = async (token0, token1, isSwapIn) => {
     try {
       const inPactCode = `(kswap.exchange.swap-exact-in
-          ${keepDecimal(token0.amount)}
-          ${keepDecimal(reduceBalance(token1.amount*(1-parseFloat(slippage)), PRECISION))}
+          (read-decimal 'token0Amount)
+          (read-decimal 'token1AmountWithSlippage)
           [${token0.address} ${token1.address}]
           ${JSON.stringify(account.account)}
           ${JSON.stringify(account.account)}
           (read-keyset 'user-ks)
         )`
       const outPactCode = `(kswap.exchange.swap-exact-out
-          ${keepDecimal(token1.amount)}
-          ${keepDecimal(reduceBalance(token0.amount*(1+parseFloat(slippage)), PRECISION))}
+          (read-decimal 'token1Amount}
+          (read-decimal 'token0AmountWithSlippage)
           [${token0.address} ${token1.address}]
           ${JSON.stringify(account.account)}
           ${JSON.stringify(account.account)}
@@ -797,7 +854,7 @@ export const PactProvider = (props) => {
             [
               account.account,
               pair.account,
-              reduceBalance(isSwapIn ? token0.amount : reduceBalance(token0.amount*(1+parseFloat(slippage)), PRECISION)),
+              reduceBalance(isSwapIn ? token0.amount : reduceBalance(token0.amount*(1+parseFloat(slippage)), tokenData[token0.coin].precision)),
             ]
           ),
         ],
@@ -805,7 +862,13 @@ export const PactProvider = (props) => {
         gasLimit: 3000,
         chainId: chainId,
         ttl: 600,
-        envData: { "user-ks": account.guard }
+        envData: {
+          "user-ks": account.guard,
+          "token0Amount": token0.amount,
+          "token1Amount": token1.amount,
+          "token0AmountWithSlippage": reduceBalance(token0.amount*(1+parseFloat(slippage)), tokenData[token0.coin].precision),
+          "token1AmountWithSlippage": reduceBalance(token1.amount*(1-parseFloat(slippage)), tokenData[token1.coin].precision)
+        }
       }
       //alert to sign tx
       walletLoading();
@@ -1017,7 +1080,8 @@ export const PactProvider = (props) => {
         setSigView,
         pw,
         setPw,
-        storeTtl
+        storeTtl,
+        tokenData
       }}
     >
       {props.children}